# fix: Resolve ESLint errors and add production optimizations