--- a/backend/package.json
+++ b/backend/package.json
@@ -53,6 +53,7 @@
     "multer": "^2.0.2",
     "nodemailer": "^6.10.1",
     "prom-client": "^15.1.3",
+    "rate-limit-redis": "^4.2.0",
     "redis": "^4.7.0",
     "sharp": "^0.33.5",
     "socket.io": "^4.8.1",